+++ src/ACTIONet/visualization/optimize.cc
@@ -9,14 +9,14 @@
 
 
 template <typename T, bool DoMove = true>
-auto optimize_layout(const T &gradient, std::vector<float> &head_embedding,
-                     std::vector<float> &tail_embedding,
+auto optimize_layout(const T &gradient, std::vector<double> &head_embedding,
+                     std::vector<double> &tail_embedding,
                      const std::vector<unsigned int> &positive_head,
                      const std::vector<unsigned int> &positive_tail,
                      unsigned int n_epochs, unsigned int n_vertices,
-                     const std::vector<float> &epochs_per_sample,
-                     float initial_alpha, float negative_sample_rate,
-                     std::size_t n_threads = 0, std::size_t grain_size = 1, unsigned int seed = 0) -> std::vector<float> {
+                     const std::vector<double> &epochs_per_sample,
+                     double initial_alpha, double negative_sample_rate,
+                     std::size_t n_threads = 0, std::size_t grain_size = 1, unsigned int seed = 0) -> std::vector<double> {
   uwot::Sampler sampler(epochs_per_sample, negative_sample_rate);
 
   uwot::SgdWorker<T, DoMove> worker(
@@ -24,7 +24,7 @@ auto optimize_layout(const T &gradient, std::vector<float> &head_embedding,
       tail_embedding, head_embedding.size() / n_vertices, seed);
 
   const auto n_epochs_per_sample = epochs_per_sample.size();
-  float alpha = initial_alpha;
+  double alpha = initial_alpha;
 	
   for (auto n = 0U; n < n_epochs; n++) {	  
     worker.set_alpha(alpha);
@@ -35,24 +35,24 @@ auto optimize_layout(const T &gradient, std::vector<float> &head_embedding,
     } else {
       worker(0, n_epochs_per_sample);
     }
-    alpha = initial_alpha * (1.0 - (float(n) / float(n_epochs)));
+    alpha = initial_alpha * (1.0 - (double(n) / double(n_epochs)));
   }
   
   return head_embedding;
 }
 
-std::vector<float>  optimize_layout_umap(
-    std::vector<float> head_vec, std::vector<float> tail_vec,
+std::vector<double>  optimize_layout_umap(
+    std::vector<double> head_vec, std::vector<double> tail_vec,
     const std::vector<unsigned int> positive_head,
     const std::vector<unsigned int> positive_tail, unsigned int n_epochs,
-    unsigned int n_vertices, const std::vector<float> epochs_per_sample,
-    float a, float b, float gamma, float initial_alpha,
-    float negative_sample_rate, bool approx_pow,
+    unsigned int n_vertices, const std::vector<double> epochs_per_sample,
+    double a, double b, double gamma, double initial_alpha,
+    double negative_sample_rate, bool approx_pow,
     std::size_t n_threads = 0, std::size_t grain_size = 1,
     bool move_other = true) {
 
 
-  std::vector<float> result;
+  std::vector<double> result;
   if (approx_pow) {
     const uwot::apumap_gradient gradient(a, b, gamma);
     if (move_other) {
@@ -85,16 +85,16 @@ std::vector<float>  optimize_layout_umap(
 
 }
 
-std::vector<float> optimize_layout_tumap(
-    std::vector<float> head_vec, std::vector<float> tail_vec,
+std::vector<double> optimize_layout_tumap(
+    std::vector<double> head_vec, std::vector<double> tail_vec,
     const std::vector<unsigned int> positive_head,
     const std::vector<unsigned int> positive_tail, unsigned int n_epochs,
-    unsigned int n_vertices, const std::vector<float> epochs_per_sample,
-    float initial_alpha, float negative_sample_rate,
+    unsigned int n_vertices, const std::vector<double> epochs_per_sample,
+    double initial_alpha, double negative_sample_rate,
     std::size_t n_threads = 0, std::size_t grain_size = 1,
     bool move_other = true) {
 		
-  std::vector<float> result;
+  std::vector<double> result;
   const uwot::tumap_gradient gradient;
 
   if (move_other) {
@@ -113,16 +113,16 @@ std::vector<float> optimize_layout_tumap(
 
 }
 
-std::vector<float>  optimize_layout_largevis(
-    std::vector<float> head_vec, const std::vector<unsigned int> positive_head,
+std::vector<double>  optimize_layout_largevis(
+    std::vector<double> head_vec, const std::vector<unsigned int> positive_head,
     const std::vector<unsigned int> positive_tail, unsigned int n_epochs,
-    unsigned int n_vertices, const std::vector<float> epochs_per_sample,
-    float gamma, float initial_alpha, float negative_sample_rate,
+    unsigned int n_vertices, const std::vector<double> epochs_per_sample,
+    double gamma, double initial_alpha, double negative_sample_rate,
     std::size_t n_threads = 0, std::size_t grain_size = 1) {
 
   const uwot::largevis_gradient gradient(gamma);
 
-  std::vector<float> result = optimize_layout<uwot::largevis_gradient, true>(
+  std::vector<double> result = optimize_layout<uwot::largevis_gradient, true>(
         gradient, head_vec, head_vec, positive_head, positive_tail, n_epochs,
         n_vertices, epochs_per_sample, initial_alpha, negative_sample_rate,
         n_threads, grain_size);
